Possible fix for Bug #60 - segfault when using -s
I am not sure what was the cause of the segfault, but it crashes
when trying to clone the unloader for the second thread. I think
that somehow the primary_unloader gets corrupted when a new unloader
is installed into the primary thread. By cloning the unloader before
starting the loop, it prevents this happening. I can't see any side
effects...